Move `ttt` to app dir
Temporary `use client` workarounds
