+++ app/(ttt)/TicTacToeBoard.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import {ReactNode, useEffect} from 'react';
