+++ app/(ttt)/TicTacToeScoreIndicator.tsx
