+++ app/(ttt)/UltimateTicTacToeBoard.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import TicTacToeBoard, {
     BoardStatus,
     defaultTTTBoard,
